Extract Burn modal
